NN-2936 adjudication history part2
