+++ backend/routes.js
@@ -37,7 +37,6 @@ const cellMoveRouter = require('./routes/cellMoveRouter')
 
 const videolinkPrisonerSearchController = require('./controllers/videolink/search/videolinkPrisonerSearch')
 const amendCaseNNoteRouter = require('./routes/caseNoteAmendmentRouter')
-const adjudicationsHistoryRouter = require('./routes/adjudicationHistoryRouter')
 
 const currentUser = require('./middleware/currentUser')
 const systemOauthClient = require('./api/systemOauthClient')
@@ -219,8 +218,6 @@ const setup = ({
     '/prisoner/:offenderNo/case-notes/amend-case-note/:caseNoteId',
     amendCaseNNoteRouter({ elite2Api, caseNotesApi, logError })
   )
-
-  router.use('/offenders/:offenderNo/adjudications', adjudicationsHistoryRouter(elite2Api, logError))
 
   return router
 }
