+++ backend/routes/prisonerProfileRouter.js
@@ -13,11 +13,14 @@ const prisonerProfessionalContacts = require('../controllers/prisonerProfile/pri
 const prisonerCellHistory = require('../controllers/prisonerProfile/prisonerCellHistory')
 const prisonerLocationHistory = require('../controllers/prisonerProfile/prisonerLocationHistory')
 const prisonerAdjudicationDetails = require('../controllers/prisonerProfile/prisonerAdjudicationDetails')
+const adjudicationsController = require('../controllers/prisonerProfile/adjudicationHistory')
 
 const prisonerProfileServiceFactory = require('../services/prisonerProfileService')
 const personServiceFactory = require('../services/personService')
 const paginationService = require('../services/paginationService')
 const referenceCodesServiceFactory = require('../controllers/reference-codes-service')
+
+const adjudicationsHistoryService = require('../services/adjudicationHistory')
 
 const router = express.Router({ mergeParams: true })
 
@@ -45,6 +48,7 @@ const controller = ({
   })
   const personService = personServiceFactory(elite2Api)
   const referenceCodesService = referenceCodesServiceFactory(elite2Api)
+  const adjudicationHistoryService = adjudicationsHistoryService(elite2Api)
 
   router.get('/', prisonerQuickLook({ prisonerProfileService, elite2Api, logError }))
   router.get('/image', prisonerFullImage({ elite2Api, logError }))
@@ -75,6 +79,16 @@ const controller = ({
   router.get('/location-history', prisonerLocationHistory({ elite2Api, logError }))
 
   router.get('/adjudications/:adjudicationNumber', prisonerAdjudicationDetails({ elite2Api, logError }))
+
+  router.use(
+    '/adjudications',
+    adjudicationsController({
+      adjudicationHistoryService,
+      paginationService,
+      elite2Api,
+      logError,
+    })
+  )
 
   return router
 }
