+++ backend/setupRedirects.js
@@ -10,5 +10,13 @@ module.exports = () => {
 
   router.get('/content/support', (req, res) => res.redirect(301, config.app.supportUrl))
 
+  router.get('/offenders/:offenderNo/adjudications', (req, res) =>
+    res.redirect(301, `/prisoner/${req.params.offenderNo}/adjudications`)
+  )
+
+  router.get(`/api/offenders/:offenderNo/adjudications/:adjudicationNumber`, (req, res) =>
+    res.redirect(301, `/prisoner/${req.params.offenderNo}/adjudications/${req.params.adjudicationNumber}`)
+  )
+
   return router
 }
