+++ views/prisonerProfile/adjudicationHistory.njk
@@ -1,4 +1,5 @@
-{% extends "./partials/layout.njk" %}
+{% extends "../partials/layout.njk" %}
+
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 {% from "govuk/components/textarea/macro.njk" import govukTextarea %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
